edemamRoute boiler plate for route and seed
