+++ controllers/index.js
@@ -1,1 +1,9 @@
-console.log('main routes')
+const router = require('express').Router();
+
+const apiRoutes = require('./api');
+const loginRoute = require('./loginRoute');
+
+router.use('/', loginRoute);
+router.use('/api', apiRoutes);
+
+module.exports = router;
