+++ seeds/edemam.js
@@ -1,0 +1,11 @@
+const sequelize = require('./config/connection');
+const { User, Recipe } = require('../models');
+
+const userData = require('./userData.json');
+const edemamData = require('./edemamRoute.js');
+
+const seedDatabase = async () => {
+ 
+};
+
+seedDatabase();
