--- seeds/edemame.js
@@ -1,1 +1,0 @@
-console.log("hello world");
